issue 9854 - extract helper to external utility class
